@@ -6,14 +6,14 @@ import java.util.Deque;
 /**
  * Einfacher Taschenrechner mit UPN
  * 
- * @author Ren� Anderes
+ * @author René Anderes
  */
 public class Calc {
 
     private Deque<BigDecimal> stack = new ArrayDeque<BigDecimal>();
     
     /**
-     * F�gt die Zahl zuoberts auf den Stack hinzu.
+     * Fügt die Zahl zuoberts auf den Stack hinzu.
      * @param bigDecimal Zahl
      */
     public void addToStack(BigDecimal bigDecimal) {
@@ -29,7 +29,7 @@ public class Calc {
     }
 
     /**
-     * L�scht den obersten Wert des Stacks
+     * Löscht den obersten Wert des Stacks
      */
     public void removeFromStack() {
         stack.remove();
